requirements.txt
torch~=1.13.1
numpy~=1.23.4
einops~=0.5.0
torchvision~=0.14.1
pandas~=1.5.1
tqdm~=4.64.1
albumentations~=1.3.0
matplotlib~=3.6.2
pillow~=9.3.0
pyyaml~=6.0
lmdb~=1.4.0
natsort~=8.2.0
timm~=0.6.11
torchmetrics~=0.10.2
setuptools~=65.5.0
scikit-learn~=1.1.3
